// src/app/types.ts
export interface CTADataProps {
  backgroundImage: string;
  title: string;
  buttonText: string;
  buttonUrl: string;
}

export interface CardDataProps {
  title: string;
  buttonLink: string;
  backgroundImage: string;
  tags: string[];
}
// ...
export interface ServiceDetailsProps {
  breadcrumb: {
    label: string;
    link: string;
    active: boolean;
  }[];
  titleHighlight: string;
  video: VideoDataProps;
  cta: CTADataProps;
  description: string;
  iconBoxes: {
    iconClass: string;
    title: string;
  }[];
}
// ...
export interface TestimonialDataProps {
  sectionTitle: string;
  testimonials: {
    text: string;
    name: string;
    designation: string;
    img: string;
  }[];
}

// Video block section
export interface VideoDataProps {
  backgroundImage: string;
  videoUrl: string;
}
// ...
export interface WorkingProcessDataProps {
  sectionTitle: string;
  subtitle: string;
  logo: string;
  steps: {
    title: string;
    description: string;
  }[];
}